move inventory table to inventory summary table
