--- src/app/motor-inventory/motor-inventory-summary/inventory-summary-overview/inventory-summary-overview.component.css
@@ -1,36 +1,0 @@
-.sticky-left{
-    position: -webkit-sticky;
-    position: sticky;
-    left: 0;
-}
-
-.table{
-    font-size: small;
-    margin-bottom: 0px;
-}
-
-
-.name-cell, .header-cell{
-    background-color: white;
-}
-
-.name-cell{
-    font-weight: bold;
-}
-
-th, td{
-    min-width: 75px;
-}
-
-.max-height{
-    max-height: 100%;
-}
-
-.sort-by-active{
-    text-decoration: underline;
-}
-
-.header-cell:hover{
-    text-decoration: underline;
-    cursor: pointer;
-}
