+++ src/app/motor-inventory/motor-inventory-summary/inventory-summary-overview/inventory-summary-overview.component.ts
@@ -1,36 +1,14 @@
-import { Component, OnInit, Input } from '@angular/core';
-import { MotorInventoryData } from '../../motor-inventory';
-import { MotorInventoryService } from '../../motor-inventory.service';
-import { InventorySummaryOverviewService, InventorySummaryData } from './inventory-summary-overview.service';
-import { Settings } from '../../../shared/models/settings';
-
+import { Component, OnInit } from '@angular/core';
 @Component({
   selector: 'app-inventory-summary-overview',
   templateUrl: './inventory-summary-overview.component.html',
   styleUrls: ['./inventory-summary-overview.component.css']
 })
 export class InventorySummaryOverviewComponent implements OnInit {
-  @Input()
-  settings: Settings;
-  
-  inventorySummaryData: InventorySummaryData;
-  sortByField: string = 'name';
-  sortByDirection: string = 'desc';
-  constructor(private motorInventoryService: MotorInventoryService, private inventorySummaryOverviewService: InventorySummaryOverviewService) { }
+
+  constructor() { }
 
   ngOnInit(): void {
-    let motorInventoryData: MotorInventoryData = this.motorInventoryService.motorInventoryData.getValue();
-    this.inventorySummaryData = this.inventorySummaryOverviewService.getInventorySummaryData(motorInventoryData, this.settings);
   }
 
-  setSortByField(str: string) {
-    if (this.sortByField == str) {
-      if (this.sortByDirection == 'desc') {
-        this.sortByDirection = 'asc';
-      } else {
-        this.sortByDirection = 'desc';
-      }
-    }
-    this.sortByField = str;
-  }
 }
