+++ src/app/motor-inventory/motor-inventory-summary/inventory-summary-table/inventory-summary-table.component.html
@@ -1,1 +1,28 @@
-<p>inventory-summary-table works!</p>
+<div class="d-flex w-100 scroll-item max-height">
+    <table class="table table-sm table-bordered table-hover">
+        <thead>
+            <tr>
+                <th class="header-cell sticky-top" *ngFor="let field of inventorySummaryData.fields; let index = index;"
+                    [ngClass]="{'sticky-left': index == 0}"><a
+                        [ngClass]="{'sort-by-active': field.value == sortByField}"
+                        (click)="setSortByField(field.value)">
+                        <span class="fa fa-chevron-up"
+                            *ngIf="sortByField == field.value && sortByDirection == 'desc'"></span>
+                        <span class="fa fa-chevron-down"
+                            *ngIf="sortByField == field.value && sortByDirection == 'asc'"></span>
+                        {{field.display}}</a>
+                </th>
+            </tr>
+        </thead>
+        <tbody>
+            <tr *ngFor="let data of inventorySummaryData.motorData | motorSummaryTable:sortByField:sortByDirection">
+                <td *ngFor="let d of data; let index = index;" [ngClass]="{'sticky-left name-cell': index == 0}">
+                    <span *ngIf="!d.pipe">{{d.value}} <span *ngIf="d.unit" [innerHTML]="d.unit"></span></span>
+                    <span *ngIf="d.pipe && d.pipe == 'motorEfficiencyClass'">{{d.value | motorEfficiencyClass}}</span>
+                    <span *ngIf="d.pipe && d.pipe == 'date'">{{d.value | date}}</span>
+                    <span *ngIf="d.pipe && d.pipe == 'currency'">{{d.value | currency}}</span>
+                </td>
+            </tr>
+        </tbody>
+    </table>
+</div>
