+++ src/app/motor-inventory/motor-inventory-summary/inventory-summary-table/inventory-summary-table.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { InventorySummaryTableService } from './inventory-summary-table.service';
+
+describe('InventorySummaryTableService', () => {
+  let service: InventorySummaryTableService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(InventorySummaryTableService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
