+++ src/app/motor-inventory/motor-inventory-summary/inventory-summary-table/motor-summary-table.pipe.spec.ts
