+++ src/app/motor-inventory/motor-inventory-summary/motor-inventory-summary.component.html
@@ -1,4 +1,4 @@
-<app-inventory-summary-overview *ngIf="summaryTab == 'overview'" [settings]="settings"></app-inventory-summary-overview>
-<app-inventory-summary-table *ngIf="summaryTab == 'table'"></app-inventory-summary-table>
+<app-inventory-summary-overview *ngIf="summaryTab == 'overview'"></app-inventory-summary-overview>
+<app-inventory-summary-table *ngIf="summaryTab == 'table'" [settings]="settings"></app-inventory-summary-table>
 <app-inventory-summary-motor-details *ngIf="summaryTab == 'motor-details'"></app-inventory-summary-motor-details>
 <app-inventory-summary-graphs *ngIf="summaryTab == 'graphs'"></app-inventory-summary-graphs>
