+++ src/app/motor-inventory/motor-inventory-summary/motor-inventory-summary.module.ts
@@ -7,7 +7,8 @@ import { InventorySummaryMotorDetailsComponent } from './inventory-summary-motor
 import { MotorInventorySummaryComponent } from './motor-inventory-summary.component';
 import { InventorySummaryOverviewService } from './inventory-summary-overview/inventory-summary-overview.service';
 import { SharedPipesModule } from '../../shared/shared-pipes/shared-pipes.module';
-import { MotorSummaryTablePipe } from './inventory-summary-overview/motor-summary-table.pipe';
+import { MotorSummaryTablePipe } from './inventory-summary-table/motor-summary-table.pipe';
+import { InventorySummaryTableService } from './inventory-summary-table/inventory-summary-table.service';
 
 @NgModule({
   declarations: [
@@ -26,7 +27,8 @@ import { MotorSummaryTablePipe } from './inventory-summary-overview/motor-summar
     MotorInventorySummaryComponent
   ],
   providers: [
-    InventorySummaryOverviewService
+    InventorySummaryOverviewService,
+    InventorySummaryTableService
   ]
 })
 export class MotorInventorySummaryModule { }
